+++ package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "react": "^19.1.0",
+    "react-dial-knob": "^1.3.0",
     "react-dom": "^19.1.0",
     "typescript": "^5.8.3"
   },
