+++ src/components/PlayRadioButton.view.tsx
@@ -2,9 +2,11 @@ import React from 'react';
 
 export const PlayRadioButton: React.FC<{
   isPlaying: boolean;
+  isLoading: boolean;
   onClick: () => void;
-}> = ({ isPlaying, onClick }) => (
+}> = ({ isPlaying, isLoading, onClick }) => (
   <button onClick={onClick} className={isPlaying ? 'playing' : ''}>
     {isPlaying ? 'Pause' : 'Play'} Reshet Bet
+    {isLoading && 'loading...'}
   </button>
 );
