+++ src/styles/elements.scss
@@ -1,3 +1,9 @@
+.header {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
 .footer {
   height: fit-content;
   border-top: 1px solid #595959;
@@ -62,29 +68,6 @@
     background-color: #28a745;
   }
 }
-
-// .volume-knob-container {
-//   width: 100px;
-//   margin-top: 10px;
-// }
-
-// #volume-knob {
-//   width: 100%;
-//   appearance: none;
-//   height: 6px;
-//   background: #ccc;
-//   border-radius: 3px;
-//   outline: none;
-// }
-
-// #volume-knob::-webkit-slider-thumb {
-//   appearance: none;
-//   height: 20px;
-//   width: 20px;
-//   border-radius: 50%;
-//   background: #333;
-//   cursor: pointer;
-// }
 
 input[type='range'] {
   transform: rotate(-90deg);
