+++ src/styles/global.scss
@@ -9,6 +9,9 @@ body {
 main {
   margin: 0 auto;
   padding: 2rem;
+  height: 40vh;
+  display: flex;
+  align-items: center;
 }
 
 h1 {
